Fixed Undefined Behavior Errors

+++ my_cond_branch_predictor.h
@@ -5,7 +5,7 @@
 
 struct SampleHist
 {
-      long long ghist;
+      uint64_t ghist;
       bool tage_pred;
       //
       SampleHist()
